Rename augur-tool to dp
