+++ package.json
@@ -6,36 +6,36 @@
   "license": "MIT",
   "main": "src/index.js",
   "bin": {
-    "dp": "./scripts/augur-tool/index.js"
+    "dp": "./scripts/dp/index.js"
   },
   "scripts": {
     "test": "mocha test/unit",
-    "lint": "eslint src test/unit scripts/augur-tool",
+    "lint": "eslint src test/unit scripts/dp",
     "coverage-save": "istanbul cover ./node_modules/mocha/bin/_mocha test/unit/",
     "coverage": "istanbul cover --include-all-sources --root ./src ./node_modules/mocha/bin/_mocha test/unit",
     "link": "npm link ethrpc ethereumjs-connect keythereum speedomatic",
     "build": "genversion src/version.js && babel ./node_modules/uuid-parse --source-root ./node_modules/uuid-parse -d ./node_modules/uuid-parse && babel ./src --source-root ./src -d ./build --copy-files && browserify ./exports.js > ./dist/augur.js && uglifyjs ./dist/augur.js > ./dist/augur.min.js",
     "preversion": "npm test",
     "version": "npm run build && git add -A dist src/version.js package.json package-lock.json",
-    "dp": "node scripts/augur-tool",
-    "upload:aura": "node scripts/augur-tool upload aura",
-    "upload:clique": "node scripts/augur-tool upload clique",
-    "upload:environment": "node scripts/augur-tool upload environment",
-    "upload:rinkeby": "node scripts/augur-tool upload rinkeby",
-    "upload": "node scripts/augur-tool upload",
-    "deploy:aura": "node scripts/augur-tool deploy aura",
-    "deploy:clique": "node scripts/augur-tool deploy clique",
-    "deploy:environment": "node scripts/augur-tool deploy environment",
-    "deploy:rinkeby": "node scripts/augur-tool deploy rinkeby",
-    "deploy": "node scripts/augur-tool deploy",
+    "dp": "node scripts/dp",
+    "upload:aura": "node scripts/dp upload aura",
+    "upload:clique": "node scripts/dp upload clique",
+    "upload:environment": "node scripts/dp upload environment",
+    "upload:rinkeby": "node scripts/dp upload rinkeby",
+    "upload": "node scripts/dp upload",
+    "deploy:aura": "node scripts/dp deploy aura",
+    "deploy:clique": "node scripts/dp deploy clique",
+    "deploy:environment": "node scripts/dp deploy environment",
+    "deploy:rinkeby": "node scripts/dp deploy rinkeby",
+    "deploy": "node scripts/dp deploy",
     "canned-markets:aura": "node scripts/canned-markets aura",
     "canned-markets:clique": "node scripts/canned-markets clique",
     "canned-markets:rinkeby": "node scripts/canned-markets rinkeby",
-    "canned-markets": "node scripts/augur-tool create-markets.sh",
-    "canned-orders:aura": "node scripts/augur-tool create-markets.sh aura orders",
-    "canned-orders:clique": "node scripts/augur-tool create-markets.sh clique orders",
-    "canned-orders:rinkeby": "node scripts/augur-tool create-markets.sh rinkeby orders",
-    "canned-orders": "node scripts/augur-tool create-orders",
+    "canned-markets": "node scripts/dp create-markets.sh",
+    "canned-orders:aura": "node scripts/dp create-markets.sh aura orders",
+    "canned-orders:clique": "node scripts/dp create-markets.sh clique orders",
+    "canned-orders:rinkeby": "node scripts/dp create-markets.sh rinkeby orders",
+    "canned-orders": "node scripts/dp create-orders",
     "release:dev": "npm version prerelease && git push && git push --tags && npm publish --tag dev",
     "release:patch": "npm version patch && git push && git push --tags && npm publish",
     "release:minor": "npm version minor && git push && git push --tags && npm publish",
