+++ scripts/dp/data/binary-order-book.json
