+++ scripts/dp/index.js
