+++ scripts/dp/lib/get-balances.js
