+++ scripts/dp/lib/get-private-key.js
