+++ scripts/dp/lib/print-transaction-status.js
