+++ scripts/helpers/create-simple-order.js
@@ -5,11 +5,11 @@
 var async = require("async");
 var chalk = require("chalk");
 var Augur = require("../../src");
-var approveAugurEternalApprovalValue = require("../augur-tool/lib/approve-augur-eternal-approval-value");
-var getPrivateKey = require("../augur-tool/lib/get-private-key");
+var approveAugurEternalApprovalValue = require("../dp/lib/approve-augur-eternal-approval-value");
+var getPrivateKey = require("../dp/lib/get-private-key");
 var connectionEndpoints = require("../connection-endpoints");
 var debugOptions = require("../debug-options");
-var createOrder = require("../augur-tool/lib/create-order");
+var createOrder = require("../dp/lib/create-order");
 
 var marketID = process.argv[2];
 var orderType = process.argv[3];
